requirements.txt
Mastodon.py>=1.8
atproto>=0.0.46
